+++ Shopping_Cart/Shopping_Cart_UI/package.json
@@ -2,12 +2,18 @@
   "name": "shopping_cart",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:52438",
   "dependencies": {
     "axios": "^0.18.0",
     "react": "^16.6.0",
     "react-dom": "^16.6.0",
+    "react-redux": "^5.1.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.0"
+    "react-scripts": "2.1.0",
+    "reactstrap": "^6.5.0",
+    "redux": "^4.0.1",
+    "redux-promise-middleware": "^5.1.1",
+    "redux-thunk": "^2.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
