+++ Shopping_Cart/Shopping_Cart_UI/src/App.js
@@ -1,14 +1,22 @@
 import React, { Component } from 'react';
 import Navagation from "./Navagation"
+// import {connect} from 'react-redux'
+// import {loginUser} from './actions/UserActions'
 
-class App extends Component {
+export default class App extends Component {
   render() {
     return (
       <div className="App">
-      <Navagation />
+      <Navagation
+      />
       </div>
     );
   }
 }
 
-export default App;
+// const mapStateToProps = state => {
+//   return {
+//     user: state.UserReducer
+//   }
+// }
+// export default connect(mapStateToProps)(App)
