+++ Shopping_Cart/Shopping_Cart_UI/src/componentCSS/HomePage.css
@@ -1,3 +1,7 @@
 .imageContainer{
     width: 30%;
 }
+
+.logout{
+    color: red
+}
