--- Shopping_Cart/Shopping_Cart_UI/src/src/api/LoginApi.js
@@ -1,10 +1,0 @@
-import axios from 'axios'
-
-export default class LoginApi {
-    static Login(data, onSuccess, onError){
-        axios.post("http://localhost:3000/api/userservice/login",
-        data,
-    ).then(onSuccess)
-    .catch(onError)
-    }
-}
