--- Shopping_Cart/Shopping_Cart_UI/src/src/api/RegisterApi.js
@@ -1,10 +1,0 @@
-import axios from 'axios'
-
-export default class RegisterApi {
-    static Register(data, onSuccess, onError){
-        axios.post("http://localhost:3000/api/userservice/register",
-        data,
-    ).then(onSuccess)
-    .catch(onError)
-    }
-}
