--- Shopping_Cart/Shopping_Cart_UI/src/src/componentCSS/HomePage.css
@@ -1,7 +1,0 @@
-.imageContainer{
-    width: 30%;
-}
-
-.logout{
-    color: red
-}
