--- Shopping_Cart/Shopping_Cart_UI/src/src/componentCSS/LoginRegister.css
@@ -1,24 +1,0 @@
-
-/*-------Login Page CSS--------*/
-.LoginHeader{
-text-align: center;
-background-color: white;
-color: deepskyblue
-}
-.LoginTable{
-    width: 35%;
-    background-color: white;
-
-}
-.LoginContainer{
-    padding-top: 300px;
-    padding-bottom: 15%;
-    background-color: deepskyblue
-}
-
-/*-------Register Page CSS--------*/
-.RegisterContainer{
-    padding-top: 250px;
-    padding-bottom: 15%;
-    background-color: deepskyblue
-}
