--- Shopping_Cart/Shopping_Cart_UI/src/src/components/common/TextInput.js
@@ -1,16 +1,0 @@
-import React from 'react'
-
-const TextInput = (props) => (
-    <React.Fragment>
-
-        <label htmlFor={props.id}>{props.label}</label>
-
-        <input className="form-control"
-            type={props.type}
-            name={props.id}
-            value={props.val}
-            onChange={props.handleChange}
-        />
-    </React.Fragment>
-)
-export default TextInput
