--- Shopping_Cart/Shopping_Cart_UI/src/src/index.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
-import {Provider} from 'react-redux'
-import Store from './Store'
-
-ReactDOM.render(
-<Provider store={Store}>
-<App />
-</Provider>,
-document.getElementById('root')
-);
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
