--- Shopping_Cart/Shopping_Cart_UI/src/src/reducers/UserReducer.js
@@ -1,15 +1,0 @@
-const UserReducer = (
-    state = {
-        isLoggedIn: false
-    }, action) => {
-debugger
-        switch(action.type){
-            case 'SET_LOGIN_STATE_FULFILLED':
-            state = {
-                isLoggedIn: true
-            }
-            break;
-        }
-    return state
-}
-export default UserReducer
